+++ types/src/main/kotlin/auth/AccountId.kt
@@ -1,4 +1,4 @@
 package auth
 
 @JvmInline
-public value class AccountId(public val int: Int)
+public value class AccountId(public val long: Long)
